requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiogram==2.25.1
    # via -r requirements.in
aiohttp==3.8.4
    # via aiogram
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
babel==2.9.1
    # via aiogram
certifi==2023.5.7
    # via aiogram
charset-normalizer==3.1.0
    # via aiohttp
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via yarl
magic-filter==1.0.9
    # via aiogram
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3
    # via babel
yarl==1.9.2
    # via aiohttp
